test(coding-agent): strengthen anthropic recovery matrix
